requirements.txt
aiofiles==23.1.0
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
bidict==0.22.1
certifi==2023.7.22
click==8.1.6
colorama==0.4.6
Django==4.2.4
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.1.2
fastapi==0.101.0
fastapi-socketio==0.0.10
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
markdown2==2.4.10
MarkupSafe==2.1.3
nicegui==1.3.7
orjson==3.9.2
pscript==0.7.7
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.15.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-engineio==4.5.1
python-multipart==0.0.6
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0.1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.22.0
vbuild==0.8.2
watchfiles==0.19.0
websockets==11.0.3
